+++ imglist.txt
@@ -1,5 +1,5 @@
-test1.jpg
-test2.jpg
-tennis_big4.jpg
-girls.jpg
-oscars.jpg
+test_imgs/test1.jpg
+test_imgs/test2.jpg
+test_imgs/tennis_big4.jpg
+test_imgs/girls.jpg
+test_imgs/oscars.jpg
